requirements.txt
pytest~=8.3.2
djangorestframework~=3.15.2
Django~=5.1
setuptools~=72.2.0